refactor: step1에서 작성한 인수테스트 dir 변경
