+++ src/test/java/kitchenpos/menu/acceptance/MenuRestAssured.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu;
+package kitchenpos.menu.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
