+++ src/test/java/kitchenpos/menugroup/acceptance/MenuGroupAcceptanceTest.java
@@ -1,19 +1,15 @@
-package kitchenpos.menugroup;
+package kitchenpos.menugroup.acceptance;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.AcceptanceTest;
-import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.menu.MenuRestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestFactory;
 
-import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.stream.Stream;
