+++ src/test/java/kitchenpos/menugroup/acceptance/MenuGroupRestAssured.java
@@ -1,4 +1,4 @@
-package kitchenpos.menugroup;
+package kitchenpos.menugroup.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
