+++ src/test/java/kitchenpos/order/acceptance/OrderAcceptanceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.order;
+package kitchenpos.order.acceptance;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
@@ -8,11 +8,10 @@ import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
-import kitchenpos.ordertable.OrderTableRestAssured;
+import kitchenpos.ordertable.acceptance.OrderTableRestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DynamicTest;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestFactory;
 
 import java.util.Arrays;
