+++ src/test/java/kitchenpos/ordertable/acceptance/OrderTableRestAssured.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable;
+package kitchenpos.ordertable.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
