+++ src/test/java/kitchenpos/product/acceptance/ProductRestAssured.java
@@ -1,4 +1,4 @@
-package kitchenpos.product;
+package kitchenpos.product.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
