+++ src/test/java/kitchenpos/tablegroup/acceptance/TableGroupRestAssured.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup;
+package kitchenpos.tablegroup.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
